Refactor k8sdriver
